Split repository.ts: extract admin audit + shortcut hits (5/N)
Fifth slice of the repository.ts carve-up (audit L14). Unlike the previous
four, these two domains are consumed by code that STAYS in repository.ts, so
they exercise the explicit-import half of the facade pattern:

- repository/adminAudit.ts — recordAdminAction, the admin_audit trail.
  autoEnrollMemberWithAudit (still in repository.ts) calls it inside its
  transaction, so repository.ts imports it explicitly.
- repository/shortcutHits.ts — the zero-model-call shortcut counters.
  usageStats (still in repository.ts) folds sumShortcutHits into its digest,
  so that is imported explicitly too.

`export *` re-exports for callers but does not bind names in the re-exporting
module's own scope — the same subtlety #804 hit with getResponseStyle and #809
hit with KNOWLEDGE_SEARCH_RELEVANCE_THRESHOLD.

One deliberate non-motion change: the private `Queryable` type alias
(Pick<PoolClient, 'query'>) moves to repository/shared.ts and is now EXPORTED,
because recordAdminAction takes it and two functions still in repository.ts do
too. Sharing a type alias is exactly what shared.ts is for; the alternative was
duplicating the declaration, which this campaign exists to reduce. The alias
itself is unchanged — only its location and visibility. Types only, so no
runtime surface changes.

All three moved functions verified byte-identical against origin/main.
repository.ts is now 6,135 lines, down from 7,137 at the start of the campaign.
